Rename to SciJava Ops Indexer
@ctrueden and @hinerm and I decided that we should specialize this
module on Ops, because we don't yet have a use case beyond Ops.
