+++ pom.xml
@@ -56,11 +56,11 @@
 		<module>scijava/scijava-discovery</module>
 		<module>scijava/scijava-discovery-test</module>
 		<module>scijava/scijava-function</module>
-		<module>scijava/scijava-javadoc-parser</module>
 		<module>scijava/scijava-log2</module>
 		<module>scijava/scijava-meta</module>
 		<module>scijava/scijava-ops-api</module>
 		<module>scijava/scijava-ops-engine</module>
+		<module>scijava/scijava-ops-indexer</module>
 		<module>scijava/scijava-ops-spi</module>
 		<module>scijava/scijava-parse2</module>
 		<module>scijava/scijava-persist</module>
