--- scijava/scijava-javadoc-parser/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,0 @@
-org.scijava.javadoc.parser.JavadocAnnotationProcessor
