+++ scijava/scijava-ops-indexer/.gitignore
