+++ scijava/scijava-ops-indexer/LICENSE.txt
