+++ scijava/scijava-ops-indexer/pom.xml
@@ -9,10 +9,10 @@
 		<relativePath>../..</relativePath>
 	</parent>
 
-	<artifactId>scijava-javadoc-parser</artifactId>
+	<artifactId>scijava-ops-indexer</artifactId>
 
-	<name>SciJava Javadoc Parser</name>
-	<description>SciJava Discovery Therapi: A therapi-aided Discovery implementation</description>
+	<name>SciJava Ops Indexer</name>
+	<description>SciJava Ops Indexer: An annotation processor for indexing Ops</description>
 	<url>https://github.com/scijava/scijava</url>
 	<inceptionYear>2021</inceptionYear>
 	<organization>
@@ -92,7 +92,7 @@
 	</ciManagement>
 
 	<properties>
-		<package-name>org.scijava.javadoc.parser</package-name>
+		<package-name>org.scijava.ops.indexer</package-name>
 
 		<license.licenseName>bsd_2</license.licenseName>
 		<license.copyrightOwners>SciJava developers.</license.copyrightOwners>
