+++ scijava/scijava-ops-indexer/src/main/java/org/scijava/ops/indexer/ImplFieldData.java
@@ -1,7 +1,4 @@
-package org.scijava.javadoc.parser;
-
-import static org.scijava.javadoc.parser.RuntimeJavadocHelper.blockSeparator;
-import static org.scijava.javadoc.parser.RuntimeJavadocHelper.tagElementSeparator;
+package org.scijava.ops.indexer;
 
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
@@ -12,9 +9,8 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 import javax.lang.model.element.Element;
-import javax.lang.model.element.ElementKind;
 
-public class ImplClassData implements ImplData {
+public class ImplFieldData implements ImplData {
 	private final Map<String, Object> implNotes = new HashMap<>();
 
 	private final List<String> authors = new ArrayList<>();
@@ -25,17 +21,18 @@ public class ImplClassData implements ImplData {
 
 	private final String source;
 
+	private final String type;
 	private String description;
 
 
 
-	public ImplClassData(Element source, String doc) {
+	public ImplFieldData(Element source, String doc) {
 		if (!doc.contains("@implNote")){
 			throw new IllegalArgumentException(source + "'s Javadoc does not contain an '@implNote' tag!");
 		}
-		String[] sections = blockSeparator.split(doc);
+		String[] sections = RuntimeJavadocHelper.blockSeparator.split(doc);
 		for (String section: sections) {
-			String[] elements = tagElementSeparator.split(section);
+			String[] elements = RuntimeJavadocHelper.tagElementSeparator.split(section);
 			switch(elements[0]) {
 				case "input":
 					params.add(new ParameterTagData(ParameterTagData.IO_TYPE.INPUT, section));
@@ -50,7 +47,7 @@ public class ImplClassData implements ImplData {
 					params.add(new ParameterTagData(ParameterTagData.IO_TYPE.MUTABLE, section));
 					break;
 				case "author":
-					authors.add(tagElementSeparator.split(section, 2)[1]);
+					authors.add(RuntimeJavadocHelper.tagElementSeparator.split(section, 2)[1]);
 					break;
 				case "implNote" :
 						implType = elements[1];
@@ -77,15 +74,10 @@ public class ImplClassData implements ImplData {
 			}
 
 		}
-		// handle inner classes
-		String srcString = source.toString();
-		Element parent = source.getEnclosingElement();
-		while(parent.getKind() == ElementKind.CLASS) {
-			int badPeriod = srcString.lastIndexOf('.');
-			srcString = srcString.substring(0, badPeriod) + '$' + srcString.substring(badPeriod+1);
-			parent = parent.getEnclosingElement();
-		}
-		this.source = "javaClass:/" + URLEncoder.encode(srcString, StandardCharsets.UTF_8);
+
+		this.source = "javaField:/" + URLEncoder.encode(source.getEnclosingElement() + "$" + source, StandardCharsets.UTF_8);
+		this.type = source.asType().toString();
+
 	}
 
 	@Override public String type() {
@@ -100,6 +92,7 @@ public class ImplClassData implements ImplData {
 		Map<String, Object> map = new HashMap<>();
 		map.put("description", description);
 		map.put("authors", authors);
+		map.put("type", type);
 		map.put("parameters", params.stream().map(ParameterTagData::data).collect(
 				Collectors.toList()));
 		map.putAll(implNotes);
